% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimpleClass_function.R
\name{Array2Mat}
\alias{Array2Mat}
\title{Simple function for conversion of an array to a matrix}
\usage{
Array2Mat(Array)
}
\arguments{
\item{Array}{is an array of landmark data}
}
\value{
This function returns an array
}
\description{
This function takes an array and converts it to a matrix where rows
specimens and columns are alternating landmark variables
(e.g. X1, X2, Y1, Y2...)
}
\author{
Ardern Hulme-Beaman
}
\keyword{data}
\keyword{format}
