% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary_functions.R
\name{BoundaryCalculation}
\alias{BoundaryCalculation}
\title{Internal function for identifies boundaries from spatial correlation scores.}
\usage{
BoundaryCalculation(MapMatrix, PlotValCor)
}
\arguments{
\item{MapMatrix}{is a matrix of correlation values that correspond to grid reference locations}

\item{PlotValCor}{numeric correlation value that is used to determine the most likely origin of the specimen. This value can be calculated by using the correct cross-validation method and identifying the correlation value that will correctly identify a desired percentage of specimens (e.g. 95\%).}
}
\value{
This function returns a matrix of logical TRUE/FALSE values identifying the boundary of the desired correlation value.
}
\description{
This function takes the correlation values calculated across the spatial grid and identifies which gird points are at the boundary of of the provenancing region.
These results then allows for the construction of compiled trait boundaries from spatial provenancing excercises.
}
\author{
Ardern Hulme-Beaman
}
\keyword{Boundary}
\keyword{Spatial}
\keyword{finder;}
\keyword{internal.}
\keyword{provenancing;}
