% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary_functions.R
\name{BoundaryFinder}
\alias{BoundaryFinder}
\title{Returns geographic boundaries identified from trait distances}
\usage{
BoundaryFinder(Lat.Longs = data.frame(Lats = c(), Longs = c()),
  Ref.Dist.mat = matrix(), LongRange = c(0, 0), LatRange = c(0, 0),
  Range.Samp = 10, print.prog = TRUE, plot.Val.cor = as.numeric(X),
  Expand.map = c(Lat, Long), datadump = NA, startpoint = 1)
}
\arguments{
\item{Lat.Longs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{Ref.Dist.mat}{is a square matrix of pairwise distances among all reference specimens}

\item{LongRange}{is a vector of two elements defining the maximum and minimum Longitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{LatRange}{is a vector of two elements defining the maximum and minimum Latitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{Range.Samp}{is an integer vector of one or two elements that defines the resolution of spatial sampling. If one element is provided then both the latitude and longitude ranges are equally and evenly sampled using this value. If two elements are provided they should be in the order of latitude longitude and each range will be evenly sampled with its respective value.}

\item{print.prog}{logical whether or not to print a progress bar. Default set to FALSE.}

\item{plot.Val.cor}{numeric correlation value that is used to determine the most likely origin of the specimen. This value can be calculated by using the correct cross-validation method and identifying the correlation value that will correctly identify a desired percentage of specimens (e.g. 95\%).}
}
\value{
If verbose is TRUE then a dataframe of all values for every sampled grid reference is returned. If verbose is FALSE then only those grid references with the highest correlation values are returned.
}
\description{
This function takes the distances among all reference specimens and uses a process similar to
leave-one-out correct cross-validation to identify likely spatial trait boundaries. This method
only allows for the inclusion of specimens with specific known locations (i.e. specific
latitude-longitude coordinates). This functions requires a distance input, which allows the user
to input any desired dissimilarity or distance metrics as appropriate for the data.
}
\details{
The map plotting of this function makes use of the functions of the \code{maps} package.
Spatial polygons were constructed using functions from the \code{sp} package and calculations of polygon intersects and areas were carried out using functions from the \code{raster} package.
When the \code{print.prog} is set to TRUE, the \code{progress} function of the \code{svMisc} package is used.
}
\section{Citations}{


Original S code by Richard A. Becker, Allan R. Wilks. R version by Ray Brownrigg.
Enhancements by Thomas P Minka and Alex Deckmyn. (2017). maps: Draw Geographical Maps. R
package version 3.2.0. https://CRAN.R-project.org/package=maps

Grosjean, Ph. (2016). svMisc: SciViews-R. UMONS, Mons, Belgium.
http://www.sciviews.org/SciViews-R.

Pebesma, E.J., R.S. Bivand, 2005. sp: Classes and methods for spatial data in R. R News 5 (2)
http://cran.r-project.org/doc/Rnews/.

Robert J. Hijmans (2016). raster: Geographic Data Analysis and Modeling. R package version 2.5-8.
https://CRAN.R-project.org/package=raster
}

\author{
Ardern Hulme-Beaman
}
\keyword{Spatial}
\keyword{identification}
\keyword{provenancing}
