% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimpleClass_function.R
\name{chr2nu}
\alias{chr2nu}
\title{Internal function for quick class change of a vector to numeric}
\usage{
chr2nu(X)
}
\arguments{
\item{x}{is a vector to be converted to a numeric vector}
}
\value{
This function returns a numeric vector
}
\description{
This function takes a factor vector and changes it to a numeric.
This is primarily to avoid instances where a csv file has been opened
without column classes being specified, which can result in numeric
values being accidently read as factors, which in turn when treated
with as.numeric() will accidently be given integer values in the order
of the factor levels. By treating the values as characters first it avoids
this potential error.
}
\author{
Ardern Hulme-Beaman
}
\keyword{internal}
