% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geographic_distance_functions.R
\name{Geo.Dist2Point}
\alias{Geo.Dist2Point}
\title{Returns vector of Haversine dists}
\usage{
Geo.Dist2Point(RefLatLongs = data.frame(Lats, Longs),
  TargetLatLong = c(Lat, Long))
}
\arguments{
\item{RefLatLongs}{a matrix of n rows by two columns where n is the number of coordinates and the columns are Latitude and Longitude values in that order.}

\item{TargetLatLong}{a vector of two elements that are Latitude and Longitude values in that order. This is the coordinate of interest which all distances will be calculated to.}
}
\value{
The distance in metres between the point of interest (TargetLatLong) to all other points in the reference material (RefLatLongs).
}
\description{
This function uses the \code{geosphere} package function \code{distHaversine} to calculate multiple distances between a set of Latitude Longitude coordinates to a target Latitude Longitude coordinate.
}
\section{Citations}{


Sinnott, R.W, 1984. Virtues of the Haversine. Sky and Telescope 68(2): 159

Robert J. Hijmans (2017). geosphere: Spherical Trigonometry. R package version 1.5-7.
https://CRAN.R-project.org/package=geosphere
}

\author{
Ardern Hulme-Beaman
}
\keyword{Haversine}
\keyword{distance}
\keyword{geographic}
\keyword{internal}
