% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geographic_distance_functions.R
\name{Geo.Dist2PointTable}
\alias{Geo.Dist2PointTable}
\title{Returns a table of Haversine dists}
\usage{
Geo.Dist2PointTable(RefLatLongs = data.frame(Lats, Longs), IDs = NA)
}
\arguments{
\item{RefLatLongs}{a matrix of n rows by two columns where n is the number of coordinates and the columns are Latitude and Longitude values in that order.}

\item{IDs}{a vector of unique IDs that correspond with (and are in the same order as) the Latitude Longitude coordinates. These IDs will then be used to name the columns and the rows of the returned table of pairwise distances. The default is set to NA. If set to NA the returned table will not have named columns or rows.}
}
\value{
A square matrix of pairwise distances in metres among all inputted coordinates (RefLatLongs).
}
\description{
This function uses the \code{geosphere} package function \code{distHaversine} to calculate pairwise distances among all inputted Latitude Longitude coordinates.
}
\section{Citations}{


Sinnott, R.W, 1984. Virtues of the Haversine. Sky and Telescope 68(2): 159

Robert J. Hijmans (2017). geosphere: Spherical Trigonometry. R package version 1.5-7.
https://CRAN.R-project.org/package=geosphere
}

\author{
Ardern Hulme-Beaman
}
\keyword{Haversine}
\keyword{distances}
\keyword{geographic}
