% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoOrigins-package.R
\docType{package}
\name{GeoOrigins}
\alias{GeoOrigins}
\title{GeoOrigins: functions for spatial exploration of trait boundaries and specimen provenance}
\description{
GeoOrigins provides functions for provenancing specimens of unknown geographic origin to spatial regions. It also
allows for the exploration and identification of trait boundaries from a distance matrix. The package includes
three datasets, which are worked through in an accompanying vignette.
}
\details{
For more information, updates or queires please see:
\url{https://github.com/ArdernHB/GeoOrigins}
}
\section{GeoOrigins functions}{


BoundaryCalculation

BoundaryFinder

BoundaryFinderPar

GeoDist2PointTable

IDbyDistanceDistInput

IDbyDistanceDistInputCCV

IDbyDistanceDistInputCCVPar

IDbyDistanceRawData

IDbyDistanceRawDataCCV

IDbyDistanceRawDataCCVPar

PlotBoundaries

ProcDistanceTable

ProcDistanceTablePar
}

\section{Datasets}{

Rpraetor: geometric morphometric dataset of tooth outlines of Rattus praetor (Large New Guinea spiny rat)

Marvalis: geometric morphometric dataset of tooth outlines of Microtus arvalis (Common vole)

Fteydea: Song distance dataset of Fingilla teydae (Tenerifian Blue Chaffinch)
}

\section{Dataset sources}{

Hulme-Beaman, A., Cucchi, T., Evin, A., Searle, J. B., & Dobney, K. (2018). Exploring Rattus praetor (Rodentia, Muridae) as a possible species complex using geometric morphometrics on dental morphology. Mammalian Biology, 92(1), 62-67.
https://doi.org/10.1016/j.mambio.2018.04.002

Cucchi, T., Barnett, R., Martinkova, N., Renaud, S., Renvoise, E., Evin, A., ... & Quere, J. P. (2014). The changing pace of insular life: 5000 years of microevolution in the Orkney vole (Microtus arvalis orcadensis). Evolution, 68(10), 2804-2820.
https://doi.org/10.1111/evo.12476
}

\section{Maintainer}{

Ardern Hulme-Beaman <ardernhb@gmail.com>
}

\author{
Ardern Hulme-Beaman
Anna Rudzinski
}
