% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProvDistInput_function.R
\name{IDbyDistance.DistInput.CCV}
\alias{IDbyDistance.DistInput.CCV}
\title{Spatial Provenancing Correct Cross-Validation calculation from distance data}
\usage{
IDbyDistance.DistInput.CCV(Lat.Longs, Dist.data.mat, verbose = TRUE,
  print.prog = TRUE, Prov.Confidence = 0.95)
}
\arguments{
\item{Lat.Longs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{Dist.data.mat}{is a square matrix of pairwise distances among all reference specimens}

\item{verbose}{logical whether or not a matrix of spatial correlation values is returned or not. Default is set to TRUE.}

\item{print.prog}{logical whether or not to print a progress bar. Default set to FALSE.}

\item{Prov.Confidence}{is a value between 0 and 1 indicating the confidence level that is desired for spatial provenancing.}
}
\value{
If verbose is set to FALSE then a list of a single object containing the correlation value at the required confidence interval is returned. If verbose is set to TRUE then a list is returned with two objects: the first is the correlation value at the required confidence interval; the second a dataframe of coordinates and the spatial-trait correlation values at the true locations of each specimen.
 @details This method also makes use of the \code{cor.test} function from the \code{stats} package. When the \code{print.prog} is set to TRUE, the \code{progress} function of the \code{svMisc} package is used.
The map plotting of this function makes use of the functions of the \code{maps} package.
}
\description{
This function takes pairwise distances among all reference specimens with known spatial origins
and uses those distances to calculate the correllation value that would be required to
correctly provenancing them in a test. This is achieved by a leave-one-out procedure. If a
cross-validation procedure that removes more than one specimen from the reference dataset is
desired then it is recommended that the validate function of the IDbyDistance.DistInput method be used
with the validate argument in a loop.
}
\section{Citations}{


Original S code by Richard A. Becker, Allan R. Wilks. R version by Ray Brownrigg.
Enhancements by Thomas P Minka and Alex Deckmyn. (2017). maps: Draw Geographical Maps. R
package version 3.2.0. https://CRAN.R-project.org/package=maps

Grosjean, Ph. (2016). svMisc: SciViews-R. UMONS, Mons, Belgium.
http://www.sciviews.org/SciViews-R.
}

\author{
Ardern Hulme-Beaman
}
