% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelProcessing_functions.R
\name{IDbyDistance.DistInput.Par}
\alias{IDbyDistance.DistInput.Par}
\title{Spatially provenance a specimen from distance data adapted for use in parallel processing}
\usage{
IDbyDistance.DistInput.Par(Lat.Longs.Par, Dist.data.vec.Par, LongRange.Par,
  LatRange.Par, Range.Samp = 10, Validate = FALSE)
}
\arguments{
\item{Lat.Longs.Par}{is a}

\item{Dist.data.vec.Par}{is a vector of distances from each reference specimen to the specimen of interest (either a specimen with unknown provenance or another reference specimen that the user is interested in validating the provenance of)}

\item{Range.Samp}{is an integer vector of 1 or 2 elements that defines the resolution of spatial sampling. If one element is provided then both the latitude and longitude ranges are equally and evenly sampled using this value. If 2 elements are provided they should be in the order of latitude longitude and each range will be evenly sampled with its respective value.}

\item{Validate}{logical whether or not to run a correct cross-validation analysis to find the lowest required correlation value for correct identification.}
}
\value{
If verbose is TRUE then a dataframe of all values for every sampled grid reference is returned. If verbose is FALSE then only those grid references with the highest correlation values are returned.
}
\description{
This function takes the distances from an unknown specimen to all reference specimens
and uses these distances to calculate a likely spatial provenance. Note that this
procedured can only be applied to one unknown specimen at a time. This distance input function
allows the user to input any dissimilarity or distance desired and as appropriate for the data.
The function has two applications either: calculating a specimens' provenance, or alternatively
it can be used to calculate the minimum correlation coefficient needed to correctly identify a
known specimen at its true collection location. The second application of this function can work
as a correct cross-validation process if looped, but see IDbyDistance.DistInput.CCV function which
does this automatically in a leave-one-out process. However, if a corss-validation process that
removed more than one specimen from the reference dataset at a time is required then it is advised
that this be applied using this function.
}
\details{
This method also makes use of the \code{cor.test} function from the \code{stats} package. When the \code{print.prog} is set to TRUE, the \code{progress} function of the \code{svMisc} package is used.
The map plotting of this function makes use of the functions of the \code{maps} package.
}
\author{
Ardern Hulme-Beaman
}
\keyword{Spatial}
\keyword{identification}
\keyword{internal}
\keyword{provenancing}
