% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProvDistInput_function.R
\name{IDbyDistance.DistInput}
\alias{IDbyDistance.DistInput}
\title{Returns the geographic provenance of a specimen}
\usage{
IDbyDistance.DistInput(Lat.Longs = data.frame(Lats = c(), Longs = c()),
  Dist.data.vec = Dist, LongRange = c(0, 0), LatRange = c(0, 0),
  Range.Samp = 10, verbose = TRUE, print.prog = FALSE,
  Validate = FALSE, Valid.LatLongs = c(Lat, Long), Plot.Res = TRUE,
  HeatHue = c(0.1, 1), Tile.Size = 2, plot.Prov = FALSE,
  plot.Val.cor = as.numeric(X))
}
\arguments{
\item{Lat.Longs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{Dist.data.vec}{is a vector of distances from each reference specimen to the specimen of interest (either a specimen with unknown provenance or another reference specimen that the user is interested in validating the provenance of)}

\item{LongRange}{is a vector of two elements defining the maximum and minimum Longitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{LatRange}{is a vector of two elements defining the maximum and minimum Latitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{Range.Samp}{is an integer vector of one or two elements that defines the resolution of spatial sampling. If one element is provided then both the latitude and longitude ranges are equally and evenly sampled using this value. If two elements are provided they should be in the order of latitude longitude and each range will be evenly sampled with its respective value.}

\item{verbose}{logical whether or not a matrix of spatial correlation values is returned or not. Default is set to TRUE.}

\item{print.prog}{logical whether or not to print a progress bar. Default set to FALSE.}

\item{Validate}{logical whether or not to run a correct cross-validation analysis to find the lowest required correlation value for correct identification.}

\item{Valid.LatLongs}{if the process is carried out on a specimen of known location `(e.g. Validate=TRUE)`, then the latitude longitude coordinates for that location should be provided here in that order.}

\item{Plot.Res}{logical whether or not to plot the provenancing map with heat values of most likely spatial origin.}

\item{HeatHue}{numeric vector of two elements each between 0 and 1. The first should be the hue value on the HSV scale; the second value should be the level of transparency of the colour used.}

\item{Tile.Size}{numeric to dictate the pixel size of the heat mapping colour values.}

\item{plot.Prov}{logical if the map should be printed with a polygon demarking a contour at a user defined correlation value.}

\item{plot.Val.cor}{numeric correlation value that is used to determine the most likely origin of the specimen. This value can be calculated by using the correct cross-validation method and identifying the correlation value that will correctly identify a desired percentage of specimens (e.g. 95\%).}
}
\value{
If verbose is TRUE then a dataframe of all values for every sampled grid reference is returned. If verbose is FALSE then only those grid references with the highest correlation values are returned.
}
\description{
This function takes the distances from an unknown specimen to all reference specimens
and uses these distances to calculate a likely spatial provenance. Note that this
procedured can only be applied to one unknown specimen at a time. This distance input function
allows the user to input any dissimilarity or distance desired and as appropriate for the data.
The function has two applications either: calculating a specimens' provenance, or alternatively
it can be used to calculate the minimum correlation coefficient needed to correctly identify a
known specimen at its true collection location. The second application of this function can work
as a correct cross-validation process if looped, but see IDbyDistance.DistInput.CCV function which
does this automatically in a leave-one-out process. However, if a corss-validation process that
removed more than one specimen from the reference dataset at a time is required then it is advised
that this be applied using this function.
}
\details{
This method also makes use of the \code{cor.test} function from the \code{stats} package. When the \code{print.prog} is set to TRUE, the \code{progress} function of the \code{svMisc} package is used.
The map plotting of this function makes use of the functions of the \code{maps} package.
}
\section{Citations}{


Original S code by Richard A. Becker, Allan R. Wilks. R version by Ray Brownrigg.
Enhancements by Thomas P Minka and Alex Deckmyn. (2017). maps: Draw Geographical Maps. R
package version 3.2.0. https://CRAN.R-project.org/package=maps

Grosjean, Ph. (2016). svMisc: SciViews-R. UMONS, Mons, Belgium.
http://www.sciviews.org/SciViews-R.
}

\author{
Ardern Hulme-Beaman
}
\keyword{Spatial}
\keyword{identification}
\keyword{provenancing}
