% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProvRawData_function.R
\name{IDbyDistance.RawData.CCV}
\alias{IDbyDistance.RawData.CCV}
\title{Spatial Provenancing Correct Cross-Validation calculation from raw data}
\usage{
IDbyDistance.RawData.CCV(Lat.Longs = data.frame(Lats = c(), Longs = c()),
  Ref.Shape = matrix(rbind(c(X1, Y1, X2, Y2), c(X1, Y1, X2, Y2))),
  Shape.Data = TRUE, Shape.dim = 2, DistMethod = c("Euc", "Proc"),
  verbose = TRUE, print.prog = TRUE, Prov.Confidence = 0.95)
}
\arguments{
\item{Lat.Longs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{Ref.Shape}{is a matrix of Reference specimen data where the rows are the individual reference specimens and the columns are the variables in the same order as the Target.Shape vector.}

\item{Shape.Data}{logical indicating whether the data is geometric morphometric shape data that requires superimposition. Default set to TRUE}

\item{Shape.dim}{integer either 2 or 3 to indicate the dimensions of landmark coordinates if the data is geometric morphometric data.}

\item{DistMethod}{determines what kind of distance calculation should be used, either Euclidean "Euc" or Procrustes "Proc". If the user wishes to use another distance or dissimilarity please use the IDbyDistance.DistInput function.}

\item{verbose}{logical whether or not a matrix of spatial correlation values is returned or not. Default is set to TRUE.}

\item{print.prog}{logical whether or not to print a progress bar. Default set to FALSE.}

\item{Prov.Confidence}{is a value between 0 and 1 indicating the confidence level that is desired for spatial provenancing.}
}
\value{
If verbose is set to FALSE then a list of a single object containing the correlation value at the required confidence interval is returned. If verbose is set to TRUE then a list is returned with two objects: the first is the correlation value at the required confidence interval; the second a dataframe of coordinates and the spatial-trait correlation values at the true locations of each specimen.
}
\description{
This function takes raw variables of reference specimens with known spatial origins
and uses euclidean distances to calculate the correllation value that would be required to
correctly provenancing them in a test. This is achieved by a leave-one-out procedure. Shape variables
can be specified and if so Procrustes distances can be calculated. If a cross-validation procedure
that removes more than one specimen from the reference dataset is desired then it is recommended that
the validate function of the IDbyDistance.RawData method be used with a loop.
}
\details{
When used for shape data and for Procrustes distances this function makes use of the \code{procGPA} and \code{procdist} functions from the \code{shapes} package. When Euclidean distances are employed the \code{dist} function of the base \code{stats} package is used.
This method also makes use of the \code{cor.test} function from the \code{stats} package. When the \code{print.prog} is set to TRUE, the \code{progress} function of the \code{svMisc} package is used.
The map plotting of this function makes use of the functions of the \code{maps} package.
}
\section{Citations}{


Original S code by Richard A. Becker, Allan R. Wilks. R version by Ray Brownrigg.
Enhancements by Thomas P Minka and Alex Deckmyn. (2017). maps: Draw Geographical Maps. R
package version 3.2.0. https://CRAN.R-project.org/package=maps

Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes

Grosjean, Ph. (2016). svMisc: SciViews-R. UMONS, Mons, Belgium.
http://www.sciviews.org/SciViews-R.
}

\author{
Ardern Hulme-Beaman
}
