% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProvRawData_function.R
\name{IDbyDistance.RawData}
\alias{IDbyDistance.RawData}
\title{Returns the geographic provenance of a shape specimen}
\usage{
IDbyDistance.RawData(Lat.Longs = data.frame(Lats = c(), Longs = c()),
  Target.Shape = c(X1, Y1, X2, Y2), Ref.Shape = matrix(rbind(c(X1, Y1,
  X2, Y2), c(X1, Y1, X2, Y2))), Shape.Data = TRUE, Shape.dim = 2,
  DistMethod = c("Euc", "Proc"), LongRange = c(0, 0), LatRange = c(0,
  0), Range.Samp = 10, verbose = TRUE, print.prog = FALSE,
  Validate = FALSE, Valid.LatLongs = c(Lat, Long), Plot.Res = TRUE,
  HeatHue = c(0.1, 1), Tile.Size = 2, plot.Prov = FALSE,
  plot.Val.cor = as.numeric(X))
}
\arguments{
\item{Lat.Longs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{Target.Shape}{is a vector of superimposed geometric morphometric coordinated in the format X1, Y1, X2, Y2... etc. (or a vector of other standardised variables that can appropriately have euclidean distances calculated between it and reference variables). NB if applied to a reference collection specimen be sure to remove it from the Ref.Shape dataset.}

\item{Ref.Shape}{is a matrix of Reference specimen data where the rows are the individual reference specimens and the columns are the variables in the same order as the Target.Shape vector.}

\item{Shape.Data}{logical indicating whether the data is geometric morphometric shape data that requires superimposition. Default set to TRUE}

\item{Shape.dim}{integer either 2 or 3 to indicate the dimensions of landmark coordinates if the data is geometric morphometric data.}

\item{DistMethod}{determines what kind of distance calculation should be used, either Euclidean "Euc" or Procrustes "Proc". If the user wishes to use another distance or dissimilarity please use the IDbyDistance.DistInput function.}

\item{LongRange}{is a vector of two elements defining the maximum and minimum Longitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{LatRange}{is a vector of two elements defining the maximum and minimum Latitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{Range.Samp}{is an integer vector of one or two elements that defines the resolution of spatial sampling. If one element is provided then both the latitude and longitude ranges are equally and evenly sampled using this value. If two elements are provided they should be in the order of latitude longitude and each range will be evenly sampled with its respective value.}

\item{verbose}{logical whether or not a matrix of spatial correlation values is returned or not. Default is set to TRUE.}

\item{print.prog}{logical whether or not to print a progress bar. Default set to FALSE.}

\item{Validate}{logical whether or not to run a correct cross-validation analysis to find the lowest required correlation value for correct identification.}

\item{Valid.LatLongs}{if the process is carried out on a specimen of known location `(e.g. Validate=TRUE)`, then the latitude longitude coordinates for that location should be provided here in that order.}

\item{Plot.Res}{logical whether or not to plot the provenancing map with heat values of most likely spatial origin.}

\item{HeatHue}{numeric vector of two elements each between 0 and 1. The first should be the hue value on the HSV scale; the second value should be the level of transparency of the colour used.}

\item{Tile.Size}{numeric to dictate the pixel size of the heat mapping colour values.}

\item{plot.Prov}{logical if the map should be printed with a polygon demarking a contour at a user defined correlation value.}

\item{plot.Val.cor}{numeric correlation value that is used to determine the most likely origin of the specimen. This value can be calculated by using the correct cross-validation method and identifying the correlation value that will correctly identify a desired percentage of specimens (e.g. 95\%).}
}
\value{
If verbose is TRUE then a dataframe of all values for every sampled grid reference is returned. If verbose is FALSE then only those grid references with the highest correlation values are returned.
}
\description{
This function takes the raw variables of an unknown specimen and reference specimens
and uses euclidean distances to calculate a likely spatial provenance. Note that this
procedured can only be applied to one unknown specimen at a time. Shape variables
can be specified and if so Procrustes distances can be calculated. It has two applications
either: calculating a specimens' provenance, or alternatively it can be used to calculate the
minimum correlation coefficient needed to correctly identify a known specimen at its true
collection location. The second application of this function can work as a correct cross-
validation process if looped, but see \code{IDbyDistance.RawData.CCV} function which does this automatically
in a leave-one-out process. However, if a corss-validation process that removed more than one specimen
from the reference dataset at a time is required then it is advised that this be applied using
this function.
}
\details{
When used for shape data and for Procrustes distances this function makes use of the \code{procGPA} and \code{procdist} functions from the \code{shapes} package. When Euclidean distances are employed the \code{dist} function of the base \code{stats} package is used.
This method also makes use of the \code{cor.test} function from the \code{stats} package. When the \code{print.prog} is set to TRUE, the \code{progress} function of the \code{svMisc} package is used.
The map plotting of this function makes use of the functions of the \code{maps} package.
}
\section{Citations}{


Original S code by Richard A. Becker, Allan R. Wilks. R version by Ray Brownrigg.
Enhancements by Thomas P Minka and Alex Deckmyn. (2017). maps: Draw Geographical Maps. R
package version 3.2.0. https://CRAN.R-project.org/package=maps

Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes

Grosjean, Ph. (2016). svMisc: SciViews-R. UMONS, Mons, Belgium.
http://www.sciviews.org/SciViews-R.
}

\author{
Ardern Hulme-Beaman
}
\keyword{Spatial}
\keyword{identification}
\keyword{provenancing}
