% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_descriptions.R
\docType{data}
\name{Marvalis}
\alias{Marvalis}
\title{\emph{Microtus arvalis} shape dataset}
\format{
A list of 5 objects:
\describe{
  \item{Info}{Specimen information for \emph{M. arvalis} dataset in the form of a data.frame with 4 columns (Population, Geographic Division, Lat, Long) and 553 rows. Rownames are specimen IDs.}
  \item{Proc.Dist}{Square distance matrix with pairwise Procrustes Distances between all 553 \emph{M. arvalis} specimens.}
  \item{Orkney.Boundary}{A list including a dataframe with provenancing area (in m2) and an array of trait boundary finding correlation values using Spearman's method for all 131 Orkney specimens for inputting into the \code{PlotBoundaries} function. This is provided as an example because the \code{BoundaryFinder} function can take a long time to compute.}
  \item{Total.Boundary}{A list including a dataframe with provenancing area (in m2) and an array of trait boundary finding correlation values using Spearman's method for all 553 vole specimens for inputting into the \code{PlotBoundaries} function. This is provided as an example because the \code{BoundaryFinder} function can take a long time to compute.}
  \item{LMs}{Landmark data of \emph{M. arvalis} dental shape in the form of an array. Please cite Cucchi et al. 2014 for use of these data.}
}
}
\source{
\url{https://doi.org/10.1111/evo.12476}
}
\usage{
Marvalis
}
\description{
A geometric morphometric derived dataset from the data of Cucchi et al. 2014.
The dataset contains a data.frame for all 553 specimens, which includes relevant
information for each specimen. This dataset also contains a Procrustes distance
table calculated pairwise between all specimens. For convenience and as the trait
boundary correlation calculations in the \code{BoundaryFinder} function is time
consuming, two arrays are also provided containing the trait boundary correlation
scores for the total dataset and also the Orkney subset. The original landmark data
can be found in the supporting information of Cucchi et al. 2014.
}
\keyword{datasets}
