% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimpleClass_function.R
\name{Mat2Array}
\alias{Mat2Array}
\title{Simple function for conversion of a matrix of landmarks to an array}
\usage{
Mat2Array(Mat, LMdim)
}
\arguments{
\item{Mat}{is a matrix of landmark data}

\item{LMdim}{is the number of dimensions of the data, either 2 or 3}
}
\value{
This function returns an array
}
\description{
This function takes a matrix where rows are specimens and columns are
alternating landmark coordinate values and converts it to a 3 dimensional
array where rows are landmarks and columns correspond with the dimensions.
}
\author{
Ardern Hulme-Beaman
}
\keyword{data}
\keyword{format}
