% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{PermArraySig}
\alias{PermArraySig}
\title{Internal function: Comparison of permutation array with true values}
\usage{
PermArraySig(PermArray, TrueMatrix, percentile = NA)
}
\arguments{
\item{PermArray}{an array of the permutation correlation results where rows and columns correspond with Lats and Longs respectively and slices are permutation iterations.}

\item{TrueMatrix}{the matrix of correlation values calculated from the true values.}

\item{Sig}{a numeric value for the signficance level the t.test is to perform to.}

\item{Alt}{a character string to be passed to the alternative argument of the t.test.}
}
\value{
Matrix of logical values for grid cell locations where the significance level is met.
}
\description{
This function takes the results of the permutation correlation test and returns a matrix of logical values of whether
}
\author{
Ardern Hulme-Beaman
}
\keyword{internal}
