% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary_functions.R
\name{PlotBoundaries}
\alias{PlotBoundaries}
\title{Plots map with boundaries from data dump folder}
\usage{
PlotBoundaries(plot.Val.cor, DataDump = TRUE, Path = NA,
  RawCorArray = NA, MapLinesWd = 1, TileSize = 4, plotLong, plotLat,
  MapExpansion = c(0, 0), Lat.Longs, Ref.PtCol = "blue",
  Ref.Pch.Size = 1, BoundaryHue = c(1, 1))
}
\arguments{
\item{plot.Val.cor}{numeric correlation value that is used to determine the most likely origin of the specimen. This value can be calculated by using the correct cross-validation method and identifying the correlation value that will correctly identify a desired percentage of specimens (e.g. 95\%).}

\item{DataDump}{if set to TRUE then the function uses files stored in the datadump folder as defined by \code{PATH}. If set to FALSE an array generated from the \code{BoundaryFinder} functions needs to be provided. Default is set to TRUE.}

\item{Path}{the directory where boundary data files have been dumped from the BoundaryFinder function. Default is set to NA.}

\item{RawCorArray}{an array of correlation values for each reference specimen where columns and rows correspond with longitude and latitude and the third dimension represents each individual reference specimen. Default is set to NA.}

\item{MapLinesWd}{is a single numeric value to set the width of the map lines of landmass boundaries.}

\item{TileSize}{is a single numeric value to set the size of the pixels/tiles that will form the boundary.}

\item{plotLong}{is a vector of numeric values for all the Longitude values sampled in the boundary finding exercise. Unlike the LongRange argument of other functions where just the maximum and minimum values are specified, this vectors should contain all Longitude value expected. As a result, this numeric vector should match the column names of the data dump files.}

\item{plotLat}{is a vector of numeric values for all the Latitude values sampled in the boundary finding exercise. Unlike the LatRange argument of other functions where just the maximum and minimum values are specified, this vectors should contain all Latitude value expected. As a result, this numeric vector should match the row names of the data dump files.}

\item{MapExpansion}{is a vector of two numeric values that define an increase in plotting area of the map (but not the area for which the boundary calculations have been carried out). The first value is added to the  This therefore allows the}

\item{Lat.Longs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{Ref.PtCol}{is the colour to be used for the reference specimen location points}

\item{Ref.Pch.Size}{is a single numeric value to set the size of the points marking where reference specimens are located.}

\item{BoundaryHue}{is a vector of 2 elements. The first element sets the colour hue value on a scale of 0 to 1 for a hsv function. The second value sets a transparancy level between 0 and 1, 0 being completely transparent.}
}
\value{
matrix of logical TRUE/FALSE values identifying the boundary of the desired correlation value
}
\description{
This function plots the results of takes the BoundaryFinder function that have been dumped in
a specified folder or are stored in an array. These results are constructed from the correlation
values calculated across the spatial grid and identifies which gird points are at the boundary of
the provenancing region. These results then allows for the construction of compiled trait boundaries
from spatial provenancing excercises.
}
\details{
The map plotting of this function makes use of the functions of the \code{maps} package.
}
\section{Citations}{


Original S code by Richard A. Becker, Allan R. Wilks. R version by Ray Brownrigg.
Enhancements by Thomas P Minka and Alex Deckmyn. (2017). maps: Draw Geographical Maps. R
package version 3.2.0. https://CRAN.R-project.org/package=maps
}

\keyword{Boundary}
\keyword{Spatial}
\keyword{finder}
\keyword{plotting}
\keyword{provenancing}
