% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapProcDist_function.R
\name{ProcDistanceTable.Par}
\alias{ProcDistanceTable.Par}
\title{Wrapper for procdist function to output a distance table with parallel processing}
\usage{
ProcDistanceTable.Par(A)
}
\arguments{
\item{A}{is a three dimensional data array of landmark x landmark dimensions x specimens}
}
\value{
This function returns a square matrix of Procrustes distances, which is required for both the \code{IDbyDistance.RawData.CCV} and the \code{BoundaryFinder} functions.
}
\description{
This function builds a square matrix of pairwise procrustes distances among specimens using
the \code{procdist} function from the \code{shapes} package. This has been set up to run parallel
using the \code{foreach} package.
}
\section{Citations}{


Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes
}

