% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapProcDist_function.R
\name{ProcDistanceTable}
\alias{ProcDistanceTable}
\title{Wrapper for procdist function to output a distance table}
\usage{
ProcDistanceTable(A)
}
\arguments{
\item{A}{is a three dimensional data array of landmark x landmark dimensions x specimens}
}
\value{
This function returns a square matrix of Procrustes distances, which is required for both the \code{IDbyDistanceRawDataCCV} and the \code{BoundaryFinder} functions.
}
\description{
This function builds a square matrix of pairwise procrustes distances among specimens using
the \code{\link[shapes]{procdist}} function from the \code{shapes} package.
}
\section{Citations}{


Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes
}

\examples{
RatDistMat <- ProcDistanceTable(Rpraetor$LMs)

}
\seealso{
\code{\link{ProcDistanceTablePar}}
}
