% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_descriptions.R
\docType{data}
\name{Rpraetor}
\alias{Rpraetor}
\title{\emph{Rattus praetor} shape dataset}
\format{
A list of 2 objects:
\describe{
  \item{Lat.Long}{Latitude longitude data for \emph{R. praetor} specimens in the form of a data.frame with 2 columns (Lat, Long) and 48 rows. Rownames are specimen IDs.}
  \item{LMs}{Landmark data of \emph{R. praetor} dental shape in the form of an array. The array has 24 rows representing landmarks, with 2 columns representing XY and 48 individuals.}
}
}
\source{
\url{https://doi.org/10.1016/j.mambio.2018.04.002}
}
\usage{
Rpraetor
}
\description{
A geometric morphometric dataset containing the XY coordinates for 24 2D landmarks
defining the shape of the first mandibular molar of 48 \emph{R. praetor}
specimens with corresponding latitude and longitude information. These data cover
approximately the entire range of this species. This speices has some geographic
structure. These data come from Hulme-Beaman et al. 2019.
}
\keyword{datasets}
