% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{transpar}
\alias{transpar}
\title{Internal function: Transparent named colour}
\usage{
transpar(Colour, alpha = 100)
}
\arguments{
\item{Colour}{A colour name from colours() function which is desired in transparent form.}

\item{alpha}{The level of transparency from 1 (completely transparent) to 100 (completely opaque) that the returned colour should be.}
}
\value{
The transparent equivalent of a named colour
}
\description{
This function takes a named colour and returns the transparent equivalent
}
\author{
Ardern Hulme-Beaman
}
\keyword{internal}
