% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceOrigin_Expansion_functions.R
\name{TwoLocationTest}
\alias{TwoLocationTest}
\title{Two location permutation significance signficance test}
\usage{
TwoLocationTest(LocationA, LocationB, HeatMappingPermResults)
}
\arguments{
\item{LocationA}{a numeric vector detailing the Latitude and Longitude values of the first location the user selects.}

\item{LocationB}{a numeric vector detailing the Latitude and Longitude values of the second location the user selects.}

\item{HeatMappingPermResults}{the output from the \code{HeatMappingPerm} function. This should be a list of two objects: TrueVarCorValues and PermutationArray.}

\item{SigLevel}{a numeric value between 1 and 0 to set the required significance level for plotting a significant ploygon. Default is set to 0.05.}
}
\value{
The difference between the two locations and what level of significance this is.
}
\description{
This function takes the output from the \code{HeatMappingPerm} function and
two user selected locations and assesses whether there is a significant difference
between the identified correlation values at the two locations. The method does
not look specifically at the user defined locations, but rather looks at the nearest
grid square in the \code{HeatMappingPerm} output; this is primarily for computing
efficiency.
}
\author{
Ardern Hulme-Beaman
}
