% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{Construct_contour}
\alias{Construct_contour}
\title{Internal function: Constructing smooth contour polygon}
\usage{
Construct_contour(LatLongs)
}
\arguments{
\item{LatLongs}{A dataframe with a 'Lat' column and a 'Long' column containing all latitude-longitude co-ordinates that are to form the basis upon which the contour is formed.}
}
\value{
Contour polygon
}
\description{
This function takes the Lat Longs that form a range around which a smooth contour is required. The contour is calculated base on kernal density estimation. This function uses the \code{kde} function of the \code{ks} Kernel Smoothing package and \code{contourLines} function of \code{grDevices} package.
}
\author{
Anna Rudzinski

Ardern Hulme-Beaman
}
\keyword{contour}
\keyword{internal}
