% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimpleClass_function.R
\name{UserInputAssessment}
\alias{UserInputAssessment}
\title{Internal function for assessment of user input}
\usage{
UserInputAssessment(
  LatLongs,
  RefDistMat = "skip",
  RefData = "skip",
  DistVec = "skip",
  Method
)
}
\arguments{
\item{LatLongs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{RefDistMat}{is a square matrix of pairwise distances among all reference specimens.}

\item{RefData}{is a matrix of Reference specimen data where the rows are the individual reference specimens and the columns are the variables in the same order as the TargetData vector.}

\item{Method}{determines what kind of correlation coefficient should be used, either "Spearman" or "Pearson". Spearman's ranked correlation coefficient does not assume a linear relationship between geographic and trait distances, whereas Pearson's coefficient does.}
}
\description{
This function assesses whether user inputs are as expected and it not
the function throws an error
}
\author{
Ardern Hulme-Beaman
}
\keyword{internal}
