% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimpleClass_function.R
\name{UserInputAssessment}
\alias{UserInputAssessment}
\title{Internal function for assessment of user input}
\usage{
UserInputAssessment(
  LatLongs,
  RefDistMat = NA,
  TargetData = NA,
  LongRange,
  LatRange,
  RangeSamp,
  Method
)
}
\arguments{
\item{LatLongs}{a matrix of n rows by 2 columns where n is the number of reference specimens in your dataset and the columns are Latitude and Longitude values in that order. These latitude-longitude coordinates should be of the locations of the reference specimens.}

\item{RefDistMat}{is a square matrix of pairwise distances among all reference specimens.}

\item{TargetData}{is a vector of unknown specimen data. If it is geometric morphometric data is should be a vector of superimposed coordinated in the format X1, Y1, X2, Y2... etc. (or a vector of other standardised variables that can appropriately have euclidean distances calculated between it and reference variables). NB if applied to a reference collection specimen be sure to remove it from the RefData dataset.}

\item{LongRange}{is a vector of 2 elements defining the maximum and minimum Longitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{LatRange}{is a vector of 2 elements defining the maximum and minimum Latitude values that the provenancing method should explore. This will also define the mapping range in the final plotted output.}

\item{RangeSamp}{is an integer vector of 1 or 2 elements that defines the resolution of spatial sampling. If one element is provided then both the latitude and longitude ranges are equally and evenly sampled using this value. If 2 elements are provided they should be in the order of latitude longitude and each range will be evenly sampled with its respective value.}

\item{Method}{determines what kind of correlation coefficient should be used, either "Spearman" or "Pearson". Spearman's ranked correlation coefficient does not assume a linear relationship between geographic and trait distances, whereas Pearson's coefficient does.}
}
\description{
This function assesses whether user inputs are as expected and it not
the function throws an error
}
\author{
Ardern Hulme-Beaman
}
